+++ packages/webpack-stats-report/src/composeReport.ts
@@ -8,13 +8,13 @@ import { createReportData } from "./createReportData";
  * @param minimumIncrease - number indicating the smallest increase in bytes that must be reported
  * @returns markdown report
  */
-export function createReport(bundleStatsResults: FileDiffResults, minimumIncrease: number = 0): string {
+export function createReport(bundleStatsResults: FileDiffResults, minimumIncrease: number = 0, atMentionThreshold: number = 0): string {
   return ["## Bundle size report"]
-      .concat(createTheReports(bundleStatsResults, minimumIncrease))
+      .concat(createTheReports(bundleStatsResults, minimumIncrease, atMentionThreshold))
     .join("\n\n");
 }
 
-const createTheReports = (bundleStatsResults: FileDiffResults, minimumIncrease: number): string[] => {
+const createTheReports = (bundleStatsResults: FileDiffResults, minimumIncrease: number, atMentionThreshold: number): string[] => {
   const reportDataWithDifference = bundleStatsResults.withDifferences.map(
     createReportData
   );
@@ -23,7 +23,7 @@ const createTheReports = (bundleStatsResults: FileDiffResults, minimumIncrease: 
   const withDifferences: string[] = [
     ...reportDataWithDifference.filter((row) => row.totalDiff !== 0),
     ...reportDataNewFiles.filter((row) => row.totalDiff !== 0),
-  ].map(data => createDetailedReport(data, minimumIncrease));
+  ].map(data => createDetailedReport(data, minimumIncrease, atMentionThreshold));
 
   const withoutDifference: string = createNoChangeReport([
     ...reportDataWithDifference.filter((row) => row.totalDiff === 0),