+++ packages/webpack-stats-report/CHANGELOG.md
@@ -1,8 +1,28 @@
 # Change Log - @microsoft/webpack-stats-report
 
-This log was last generated on Mon, 08 May 2023 08:05:08 GMT and should not be manually modified.
+This log was last generated on Tue, 01 Aug 2023 12:24:50 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 1.1.1
+
+Tue, 01 Aug 2023 12:24:50 GMT
+
+### Patches
+
+- Improved report formatting (71752651+unpervertedkid@users.noreply.github.com)
+
+## 1.1.0
+
+Wed, 05 Jul 2023 15:39:16 GMT
+
+### Minor changes
+
+- at mention asset owners if chunk sizes increase above a certain threshold (lkiniu@microsoft.com)
+
+### Patches
+
+- Bump @microsoft/webpack-stats-differ to v1.1.0 (lkiniu@microsoft.com)
 
 ## 1.0.3